+++ scripts/build.sh
@@ -2,11 +2,8 @@
 my_dir="$(dirname "$0")"
 source "$my_dir/validate.sh"
 
-$my_dir/mopub-android-sdk-unity-build.sh
+$my_dir/../mopub-android-sdk-unity/gradlew -p mopub-android-sdk-unity clean assembleRelease
 validate "Android build failed, fix before continuing."
-
-$my_dir/mopub-ios-sdk-unity-build.sh
-validate "iOS build failed, fix before continuing."
 
 $my_dir/unity-export-package.sh
 validate "Exporting the package failed."