+++ unity-sample-app/Assets/MoPub/Sample/MoPubDemoGUI.cs
@@ -240,19 +240,19 @@ public class MoPubDemoGUI : MonoBehaviour
                 new MoPub.SupportedNetwork.AdMob
                 {
                     // Network adapter configuration settings (initialization).
-                    NetworkConfiguration = {
+                    NetworkConfiguration = new Dictionary<string,object> {
                         { "key1", value },
                         { "key2", value },
                     },
 
                     // Global mediation settings (per ad request).
-                    MediationSettings = {
+                    MediationSettings = new Dictionary<string,object> {
                         { "key1", value },
                         { "key2", value },
                     },
 
                     // Additional options to pass to the MoPub servers (per ad request).
-                    MoPubRequestOptions = {
+                    MoPubRequestOptions = new Dictionary<string,object> {
                         { "key1", "value" },
                         { "key2", "value" },
                     }
@@ -276,13 +276,13 @@ public class MoPubDemoGUI : MonoBehaviour
 
                     // Fill in settings and configuration options the same way as for supported networks:
 
-                    NetworkConfiguration = { ... },
+                    NetworkConfiguration = ...,
 
                 #if UNITY_IOS  // See note above.
-                    MediationSettings    = { ... },
+                    MediationSettings = ...,
                 #endif
 
-                    MoPubRequestOptions  = { ... },
+                    MoPubRequestOptions = ...,
                 }
             */
             },