+++ unity-sample-app/Assets/MoPub/Scripts/Internal/MoPubUnityEditor.cs
@@ -165,6 +165,21 @@ public class MoPubUnityEditor : MoPubBase
 
 
     /// <summary>
+    /// Inform the platform SDK of a change in the application's pause status.
+    /// <para>
+    /// For platform-specific implementations, see
+    /// MoPubAndroid.<see cref="MoPubAndroid.OnApplicationPause(bool)"/> and
+    /// MoPubiOS.<see cref="MoPubiOS.OnApplicationPause(bool)"/>.
+    /// </para>
+    /// </summary>
+    /// <param name="paused">True when pausing, false when resuming</param>
+    internal static void OnApplicationPause(bool paused)
+    {
+        // Nothing to do, since there is no platform SDK to inform.
+    }
+
+
+    /// <summary>
     /// Allow supported SDK networks to collect user information on the basis of legitimate interest.
     /// Can also be set via MoPub.<see cref="MoPub.SdkConfiguration"/> on
     /// MoPub.<see cref="MoPubUnityEditor.InitializeSdk(MoPub.SdkConfiguration)"/>
@@ -370,12 +385,12 @@ public class MoPubUnityEditor : MoPubBase
     /// Whether the interstitial ad is ready to be shown or not.
     /// <para>
     /// For platform-specific implementations, see
-    /// MoPubAndroid.<see cref="MoPubAndroid.IsInterstialReady(string)"/> and
-    /// MoPubiOS.<see cref="MoPubiOS.IsInterstialReady(string)"/>.
+    /// MoPubAndroid.<see cref="MoPubAndroid.IsInterstitialReady(string)"/> and
+    /// MoPubiOS.<see cref="MoPubiOS.IsInterstitialReady(string)"/>.
     /// </para>
     /// </summary>
     /// <param name="adUnitId">A string with the ad unit id.</param>
-    public bool IsInterstialReady(string adUnitId)
+    public static bool IsInterstitialReady(string adUnitId)
     {
         CheckAdUnitRequested(adUnitId);
         return _requestedAdUnits.Contains(adUnitId);
@@ -391,7 +406,7 @@ public class MoPubUnityEditor : MoPubBase
     /// </para>
     /// </summary>
     /// <param name="adUnitId">A string with the ad unit id.</param>
-    public void DestroyInterstitialAd(string adUnitId)
+    public static void DestroyInterstitialAd(string adUnitId)
     {
         CheckAdUnitRequested(adUnitId);
     }